+++ apps/beacon/main.cpp
@@ -36,6 +36,7 @@
 #include "beacon/beacon_service.hpp"
 #include "beacon/beacon_setup_service.hpp"
 #include "beacon/block_entropy.hpp"
+#include "beacon/create_new_certificate.hpp"
 #include "beacon/event_manager.hpp"
 
 #include <cstdint>
@@ -55,18 +56,6 @@ using ProverPtr      = std::shared_ptr<Prover>;
 using Certificate    = fetch::crypto::Prover;
 using CertificatePtr = std::shared_ptr<Certificate>;
 using Address        = fetch::muddle::Packet::Address;
-
-ProverPtr CreateNewCertificate()
-{
-  using Signer    = fetch::crypto::ECDSASigner;
-  using SignerPtr = std::shared_ptr<Signer>;
-
-  SignerPtr certificate = std::make_shared<Signer>();
-
-  certificate->GenerateKeys();
-
-  return certificate;
-}
 
 struct DummyManifestCache : public ledger::ManifestCacheInterface
 {
